third_party/WebKit/LayoutTests/editing/inserting/insert_interchange_newline_crash.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '<div>foo^</div><div>|bar</div>',
        '</div>',
    ].join(''),
    'insertHTML <br class="Apple-interchange-newline">',
    [
        '<div contenteditable>',
            '<div>foo</div><div>|bar</div>',
        '</div>',
    ].join('')),
    'interchange-newline does nothing for DIV');
</script>
